feat: 입찰 최종 승인 시 광고주에게서 인플루언서로 돈이 전달되는 기능 구현

+++ src/main/java/com/shotty/shotty/domain/balance/dto/ChangeBalanceDto.java
@@ -1,6 +1,6 @@
 package com.shotty.shotty.domain.balance.dto;
 
 public record ChangeBalanceDto (
-        Long amount
+        int amount
 ) {
 }
